rename _init_helper module to _logging
